+++ internal/applogic/domain/model/money.go
@@ -27,9 +27,8 @@ const (
 	JPY Currency = "JPY"
 )
 
-// todo: should we use int64 for amount? or float64?
 type Money struct {
-	Amount   int64
+	Amount   float64
 	Currency Currency
 }
